pyproject.toml
[project]
name = "mcp-server-weather"
version = "0.1.1"
description = "My test MCP server for weather data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    # this is the standalone FastMCP library from https://pypi.org/project/fastmcp/
    "fastmcp>=3.0.0",
    "pytest-asyncio>=1.3.0",
]
